requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic[dotenv]
requests
python-dotenv
flake8